Fix clang-tidy warnings and code quality issues
Core changes:
- Use = default for trivial constructors (ClipManager)
- Add missing override specifiers on destructors
- Mark final classes (PluginEditorWindow)
- Initialize all member variables in constructors (ClipCommands)
- Add explicit null checks in helper methods (LFOPhaseOverlay)

API improvements:
- Pass smart pointers by const reference (AudioBridge)
- Remove unused parameter from PluginWindowManager::isPluginWindowOpen
- Add move operation deletes for MidiBridge (rule of five)

Warning fixes:
- Fix sign comparison warnings (DeviceProcessor)
- Remove unused variables and lambda captures
- Cast ignored return values to (void)
- Add [[maybe_unused]] attribute where appropriate
- Add missing TypeIds.hpp include (ClipInfo)

Simplify parameter definitions in SimpleSynthPlugin to use proper
5-argument addParam signature with custom formatters.

+++ magda/daw/audio/DeviceProcessor.cpp
@@ -483,7 +483,8 @@ void ExternalPluginProcessor::syncFromDeviceInfo(const DeviceInfo& info) {
     // Sync parameter values
     if (auto* ext = getExternalPlugin()) {
         auto params = ext->getAutomatableParameters();
-        for (size_t i = 0; i < info.parameters.size() && i < params.size(); ++i) {
+        for (size_t i = 0; i < info.parameters.size() && i < static_cast<size_t>(params.size());
+             ++i) {
             if (params[i]) {
                 params[i]->setParameter(info.parameters[i].currentValue,
                                         juce::dontSendNotification);
@@ -566,7 +567,7 @@ void ExternalPluginProcessor::parameterChanged(te::AutomatableParameter& param, 
     int parameterIndex = -1;
     if (auto* ext = getExternalPlugin()) {
         auto params = ext->getAutomatableParameters();
-        for (size_t i = 0; i < params.size(); ++i) {
+        for (size_t i = 0; i < static_cast<size_t>(params.size()); ++i) {
             if (params[i] == &param) {
                 parameterIndex = static_cast<int>(i);
                 break;
